requirements.txt
boto3==1.5.8
botocore==1.8.22
docutils==0.14
jmespath==0.9.3
PyMySQL==0.8.0
python-dateutil==2.6.1
PyYAML==3.12
s3transfer==0.1.12
six==1.11.0
SQLAlchemy==1.2.0
